Redstone logic 4
Made it so that redstone is not powered by torches that are attatched to the same block as it. Along this theme also fixed a bug causing an infintie loop with a redstone ontop of a block with a torch. Also implemented redstone powering the block they were attatched to